requirements.txt
# Flask Web Framework
Flask==3.1.2
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23

# Password Hashing
bcrypt==4.1.2

# Image Processing
Pillow==11.3.0

# Testing
pytest==8.4.2
pytest-flask==1.3.0

# Development Tools
python-dotenv==1.0.0
